requirements.txt
fastapi
uvicorn[standard]
aiokafka
redis
python-dotenv
groq
openai
pydantic
pydantic-settings
